Utils: optimized some function

+++ utils.h
@@ -39,7 +39,7 @@ public:
     static void getMemoryInfo(QString &memory, float &percent);
     static void getDiskInfo(QString &disk, float &percent);
     static void getNetworkBandWidth(unsigned long long &receiveBytes, unsigned long long &sendBytes);
-    static QString formatBytes(long bytes);
+    static QString formatBytes(unsigned long long bytes);
     static quint64 getFileSize(const QString &path);
 
     static QFileInfoList getDpkgPackages();
